requirements.txt
aioredis==1.3.1
asgiref==3.3.0
astroid==2.4.0
async-timeout==3.0.1
attrs==20.2.0
Automat==20.2.0
Brotli==1.0.7
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.2.1
Django==3.1.3
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
hiredis==1.1.0
hyperlink==20.0.1
idna==2.9
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
numpy==1.18.3
pandas==1.0.3
plotly==4.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.5.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.23.0
retrying==1.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.4.1
toml==0.10.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zope.interface==5.1.2
djangorestframework==3.12.4
